SPCES: linux-firmware: do not differentiate DVM/MVM/MLD
It's quite different to differentiate the firmware files used by these 3
drivers for users, especially for newer cards which has cryptic
codenames in the filename instead of model number.

With open-coded firmware filename, the spec file is also ready for
inclusion of newer cards.

3945/4965 cards, which are supported by iwlegacy driver instead of
iwlwifi, are still a dedicated package. These cards do not have cryptic
codenames, and use a different driver, so users should be easy to know
which package they need to install.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>